feat: add location filter to filter menu
The location filter is a standard search text input with typeahead menu
generated from the input restaurant locations. A "location" is defined
as the country, state, city, or neighborhood a restaurant is located in.
We think that users will generally want to search by any of these fields
together, rather than having separate filters for each location type.

The typeahead menu is a custom-implemented component that is designed to
integrate with the Pinia store that holds the restaurant location data.
If we switch to using a remote, queryable data store (e.g. DynamoDB) in
the future, we can implement a new `TypeaheadEngine`
and keep the `TypeaheadController` interface mostly the same
(except for the constructor props).